+++ frb/surveys/sdss.py
@@ -1,7 +1,7 @@
 """ Methods related to SDSS/BOSS queries """
 
 import numpy as np
-import pdb
+from IPython import embed
 
 from astropy import units
 from astropy.coordinates import SkyCoord
@@ -17,6 +17,17 @@ from frb.surveys import surveycoord
 from frb.surveys import catalog_utils
 from frb.surveys import images
 
+# Define the data model for DES data
+photom = {}
+photom['SDSS'] = {}
+DES_bands = ['u', 'g', 'r', 'i', 'z']
+for band in DES_bands:
+    photom['SDSS']['SDSS_{:s}'.format(band)] = 'petroMag_{:s}'.format(band.lower())
+    photom['SDSS']['SDSS_{:s}_err'.format(band)] = 'petroMagErr_{:s}'.format(band.lower())
+photom['SDSS']['SDSS_ID'] = 'objid'
+photom['SDSS']['ra'] = 'ra'
+photom['SDSS']['dec'] = 'dec'
+photom['SDSS']['SDSS_field'] = 'field'
 
 class SDSS_Survey(surveycoord.SurveyCoord):
     """
@@ -61,9 +72,10 @@ class SDSS_Survey(surveycoord.SurveyCoord):
         """
         if photoobj_fields is None:
             photoobj_fs = ['ra', 'dec', 'objid', 'run', 'rerun', 'camcol', 'field']
-            mags = ['petroMag_u', 'petroMag_g', 'petroMag_r', 'petroMag_i', 'petroMag_z']
-            magsErr = ['petroMagErr_u', 'petroMagErr_g', 'petroMagErr_r', 'petroMagErr_i', 'petroMagErr_z']
-            photoobj_fields = photoobj_fs+mags+magsErr
+            mags = ['petroMag_'+band for band in DES_bands]
+            magsErr = ['petroMagErr_'+band for band in DES_bands]
+            extinct = ["extinction_"+band for band in DES_bands]
+            photoobj_fields = photoobj_fs+mags+magsErr+extinct
 
         # Call
         photom_catalog = SDSS.query_region(self.coord, radius=self.radius, timeout=timeout,
@@ -94,17 +106,41 @@ class SDSS_Survey(surveycoord.SurveyCoord):
         photz_cat = SDSS.query_sql(query,timeout=timeout)
 
         # Match em up
-        matches = catalog_utils.match_ids(photz_cat['objid'], photom_catalog['objid'], require_in_match=False)
+        if photz_cat is not None:
+            matches = catalog_utils.match_ids(photz_cat['objid'], photom_catalog['objid'], require_in_match=False)
+        else:
+            matches = -1 * np.ones(len(photom_catalog), dtype=int)
         gdz = matches > 0
         # Init
         photom_catalog['photo_z'] = -9999.
         photom_catalog['photo_zerr'] = -9999.
         # Fill
-        photom_catalog['photo_z'][matches[gdz]] = photz_cat['redshift'][np.where(gdz)]
-        photom_catalog['photo_zerr'][matches[gdz]] = photz_cat['redshift_error'][np.where(gdz)]
+        if np.any(gdz):
+            photom_catalog['photo_z'][matches[gdz]] = photz_cat['redshift'][np.where(gdz)]
+            photom_catalog['photo_zerr'][matches[gdz]] = photz_cat['redshift_error'][np.where(gdz)]
 
         # Trim down catalog
         trim_catalog = trim_down_catalog(photom_catalog, keep_photoz=True)
+
+        # Clean up
+        trim_catalog = catalog_utils.clean_cat(trim_catalog, photom['SDSS'])
+
+        # Spectral info
+        spec_fields = ['ra', 'dec', 'z', 'run2d', 'plate', 'fiberID', 'mjd', 'instrument']
+        spec_catalog = SDSS.query_region(self.coord,spectro=True, radius=self.radius,
+                                         timeout=timeout, specobj_fields=spec_fields) # Duplicates may exist
+        trim_spec_catalog = trim_down_catalog(spec_catalog)
+        # Match
+        spec_coords = SkyCoord(ra=trim_spec_catalog['ra'], dec=trim_spec_catalog['dec'], unit='deg')
+        phot_coords = SkyCoord(ra=trim_catalog['ra'], dec=trim_catalog['dec'], unit='deg')
+        idx, d2d, d3d = match_coordinates_sky(spec_coords, phot_coords, nthneighbor=1)
+        # Check
+        if np.max(d2d).to('arcsec').value > 1.5:
+            embed(header='119 of sdss')
+        # Fill me
+        zs = -1 * np.ones_like(trim_catalog['ra'].data)
+        zs[idx] = trim_spec_catalog['z']
+        trim_catalog['z_spec'] = zs
 
         # Sort by offset
         catalog = trim_catalog.copy()
@@ -197,6 +233,7 @@ def get_url(coord, imsize=30., scale=0.39612, grid=False, label=False, invert=Fa
 
 def trim_down_catalog(catalog, keep_photoz=False, cut_within=1.5*units.arcsec):
     """
+    Cut down a catalog to keep only 1 source within cut_within
 
     Args:
         catalog (astropy.table.Table):  Input source catalog
@@ -207,6 +244,9 @@ def trim_down_catalog(catalog, keep_photoz=False, cut_within=1.5*units.arcsec):
         astropy.table.Table:  Catalog trimmed down
 
     """
+    if len(catalog) == 1:
+        return catalog
+
     # All good
     keep = np.ones_like(catalog, dtype=bool)