+++ CHANGES.rst
@@ -10,3 +10,7 @@
 - Add Virgo halo
 - Simple Milky Way methods for DM
 - Add WISE to surveys
+- Scripts to build FRB, Hosts, and specDB
+- Brings 180924 files into the public light
+- Add SDSS to surveys
+- Figure methods for basic plots